perf(hdrvdp): f32 planes/intermediates — fleet-throughput precision split
Planes, pathway outputs, pyramid bands, masking intermediates, resizes,
and visibility maps move f64→f32; scalar params, LUT construction,
frequency grids, quality weights, reductions, and all public scalar
results stay f64. HdrVdpResult::{p_map,c_map} and Band::data are now
Vec<f32> (breaking).

Official-golden tolerances re-baselined to measured f32 drift (worst
over 24 cases: P_det 1.4e-5, C_max 1.0e-4 rel, P_map 1.7e-2, res.Q
7.8e-4 — diffuse quantisation noise well below JND resolution);
tests/bit_lock.rs retired — the official-golden tolerance test is the
correctness gate. 1024² probe: full 1.85→1.28 s, score() 1.02 s.

+++ crates/hdrvdp/examples/stage_probe.rs
@@ -114,7 +114,7 @@ fn main() {
         acc[5] += t.elapsed().as_secs_f64();
 
         let t = Instant::now();
-        let l_adapt: Vec<f64> = path_ref
+        let l_adapt: Vec<f32> = path_ref
             .l_adapt
             .iter()
             .zip(&path_test.l_adapt)
